requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
pip==21.2.2
numpy==1.21.5
streamlit==1.8.0
click<=8.0.4
pillow==9.1.0
keras==2.8.0
tensorflow==2.8.0